\encoding{utf8} 
\name{MisRepARMA-internal}
\alias{MisRepARMA-internal}
\alias{estimate}

\title{Internal MisRepARMA functions}
\description{Internal MisRepARMA functions}
\usage{
estimate(data, tol, p_AR, q_MA, covars=NULL, misReport="U")
}

\details{These functions are not to be called by the user}
\seealso{
\code{\link{MisRepARMA-package}}, \code{\link{fitMisRepARMA}}, \code{\link{reconstruct}}
}

\keyword{ internal }
